+++ logger/logger.js
@@ -1,14 +1,14 @@
-var Logger = function () {};
+let Logger = function () {};
 
-Logger.prototype.info = function(logText) {
+Logger.prototype.info = (logText) => {
   console.log(new Date() + 'info::::' + logText)
 };
 
-Logger.prototype.debug = function(logText) {
+Logger.prototype.debug = (logText) => {
   console.log(new Date() + 'debug::::' + logText)
 };
 
-Logger.prototype.error = function(logText) {
+Logger.prototype.error = (logText) => {
   console.log(new Date() + 'error::::' + logText)
 };
