+++ routes/routes.js
@@ -1,0 +1,6 @@
+const express = require('express')
+const apiRouter = express()
+
+apiRouter.use('/user', require('./user'))
+
+module.exports = apiRouter
